+++ lib/schleuder/plugins/resend.rb
@@ -117,22 +117,22 @@ module Schleuder
       Array(recipients).inject({}) do |hash, email|
         keys = mail.list.keys(email)
         # Exclude unusable keys.
-        keys.select! { |key| key.usable_for?(:encrypt) }
-        case keys.size
+        usable_keys = keys.select { |key| key.usable_for?(:encrypt) }
+        case usable_keys.size
         when 1
-          hash[email] = keys.first
+          hash[email] = usable_keys.first
         when 0
           if encrypted_only
             # Don't add the email to the result to exclude it from the
             # recipients.
-            add_keys_error(mail, email, keys.size)
+            add_keys_error(mail, email, usable_keys.size, keys.size)
           else
             hash[email] = ''
           end
         else
           # Always report this situation, regardless of sending or not. It's
           # bad and should be fixed.
-          add_keys_error(mail, email, keys.size)
+          add_keys_error(mail, email, usable_keys.size, keys.size)
           if ! encrypted_only
             hash[email] = ''
           end
@@ -152,8 +152,8 @@ module Schleuder
       gpg_opts
     end
 
-    def self.add_keys_error(mail, email, keys_size)
-      mail.add_pseudoheader(:error, I18n.t("plugins.resend.not_resent_no_keys", email: email, num_keys: keys_size))
+    def self.add_keys_error(mail, email, usable_keys_size, all_keys_size)
+      mail.add_pseudoheader(:error, I18n.t("plugins.resend.not_resent_no_keys", email: email, usable_keys: usable_keys_size, all_keys: all_keys_size))
     end
 
     def self.add_error_header(mail, recipients_map)